Instead of parsing size get family information from option hash
In various places, we were parsing the size to get family information. This is
not necessary, as we can get the family information from the options hash. Also
parsing by splitting the size string by dash character is not reliable, because
the size string can be in different format in the future. For example, for AWS
size string can be like `m6gd.large`.