Make cloudify_server strictly Hetzner-only
As we developed clover, we made more assumptions about the hosting
providers and coupled the code with them, specifically Hetzner.
As reported by users, things break in the cloudify_server demo script
and we don't do a great job at actively maintaining it.

This commit applies some modifications to the cloudify_server script:
- Make the script Hetzner-only
- Ask for Hetzner details to be entered in .env file beforehand
- Allow re-issue of the script on the same host, for retry attempts
- Bump the VmHost init timeout from 10m to 30m