Allocate VmStorageVolume directly in assemble for AWS resources
Right at the time of creating the entities, we need to create the
vm_storage_volume entities as well. This is mainly because we calculate
the storage size through these entities in convergence prog to see if
there is a new server that matches resource's description of target
storage size. If not, we create a new VM infinitely. Therefore, this
must be created with the vm entity so that we create only 1 instance.