Add OIDC login flow
This was tested with the example authorization server that ships
with rodauth-oauth.

This unfortunately adds a large number of gems, but that's the
current price to pay for OIDC support.

Omniauth is not setup for dynamic providers, it expects a fixed
set of providers. So currently, if you add a provider, you need
to restart/reload Clover for it to pick up the new provider.

Each provider needs its own button. This handles that by using a
separate page per provider, with a single button. So users that
want to use OIDC login will need to have a separate page that
shows the button for their provider.