Fix stale walg credentials bug for read replicas after promo
Since the access key and secret key are only set after the timeline is
created and the first state of the timeline nexus is processed. when we
switch to a new timeline, we do not have the proper credentials, yet.
Still, we do update the walg config after timeline switch and it makes
sense. However, for AWS instances, we need to be able to trigger it
again once we create the access/secret key pair. This commit handles
that.