Add the debug gem to the development group in the Gemfile
This allows you to place a binding.break call anywhere in the code,
and have the process enter a debugging session when the call is
reached.  The debug gem is one of the gems bundled with Ruby.
I found this very helpful when debugging the OIDC login flow issues
with dynamic OIDC providers.