Support connecting and disconnecting OIDC authentication support for accounts
This is based on the existing support for connecting social logins
to accounts.  Connecting needs to be handled differently, since
we don't want to show all accounts every OIDC provider that is
supported (OIDC providers should be considered internal).

As OIDC providers are named with the OidcProvider ubid, add
omniauth_provider_name helper method that returns the display name
for the provider, and use that in the flash messages.

This makes a change to the layout so it will use the @page_message
style page if logged in and @use_page_message is set. I'm not sure
whether @page_message is already used on any page shown when logged
in.  If not, we could probably remove @use_page_message and have
the presence of @page_message trigger the alternative layout.