Avoid request to jwks endpoint for every OIDC login
This is not necessary as you can trust the TLS server verification,
per OIDC spec 3.1.3.7.6. If the server is giving you a bogus ID
token, it doesn't make sense for you to trust that the same server
is giving you a valid jwks.

Additionally, we are not doing anything with the id token other
than providing it to the userinfo endpoint to get the user
information. If the id token is not valid, then the userinfo
endpoint should raise an error and then the authentication will
not be allowed.