Give explicit description to runner script service
Previously, if no description was provided, the service used the full
command as the description. Since the command includes the JIT token,
and GitHub recently increased the token length. This resulted in overly
long descriptions.

More importantly, including the token in the description is not ideal
from a security and clarity standpoint. To address this, we now assign a
default, explicit description instead.