Set cron.use_background_workers to on for pg_cron by default
Pg_cron uses client connections by default which increases the
connection count and it doesn't really work out of the box. We could
either configure using unix sockets or let pg to use background workers.
This commit configures it to use background workers that makes it work
out of the box.