Don't trust redirect_url during social login
This is only every set by us if we are connecting an existing
account to social account. In that case, we know where it should
go, so we don't need to trust it. I'm not sure there is any
security issue in trusting it, but I'm also not sure there are
no security issues in trusting it.