Support registering OIDC provider with given client_id and client_secret
This can be used when the OIDC provider provides discovery
information, but does not support dynamic anonymous client
registration. In this case, you can just provide the base url,
id, and secret, and the code can figure out the endpoints.

Update the method comment to reflect current behavior while here.