Provide the database URL explicitly in E2E tests
I unified our workflows to make similar steps into reusable actions in
the previous commit.

We started to create the base environment with the `overwrite_envrb`
rake task, which has values for `test` and `development` environments.

Since we run E2E tests in the `production` environment, we need to
provide the database URL explicitly.