Don't show connection information until the PG is running
When we give this information too early, before the DNS records are created,
users might attempt to connect to it, which would cause their OS to cache the
negative DNS response, and they would not be able to connect to the PG until
the DNS cache is cleared. In some cases, the negative response can be cached
outside of the OS (e.g. external recursive DNS servers), which would make the
situation even worse as it wouldn't be possible to flush the cache and the only
way to resolve the issue would be to wait for the cache to expire.