Set runtime token secret for E2E
This was removed during the unification changes. However, E2E tests
require it, especially for testing Ubicloud Cache.

I’ve added it back via a rake task instead of setting it directly in the
workflow. The reason is that this value is also needed in the
development environment if you want to test Ubicloud Cache locally.

This change is part of our effort to make the development environment
setup as simple and consistent as possible.