Fix 0 maintenance_window_set param bug
typecast_params.pos_int(x) returns nil if the parameter is not a
positive integer. This is problemmatic because 0 is a valid argument for
00:00 - 02:00 maintenance window. However, using pos_int, we simply
nullify the value and return success. The problem is, using pos_int is
even more problemmatic because for a value of 26, user gets a 400,
failure because the db rejects it, however, for a non positive, they
simply get 200, but nothing happens. This commit fixes this behavior by
simply leaving the error to the database constraint.