Do not show Login Methods, Change Password, or Change Email for locked domains
Locked domains are not allowed to change any of these things, so
it is best not to display them.

Previously, the account submenu view handled a fourth argument for
whether the tab is active.  This behavior dates back to
842a476f54d739c2e0cbe2d4229707c62891c8bb, when a fourth argument
was never passed. A fourth argument was eventually passed in
b1c0a2ede74d2c8a0a9c2a3bd32bb64adf6bfff4, but it hasn't been needed
since a4a563b49b03177dc3dc17d1add6c24c03dd4b46, when the /account
path was redirected to /account/multifactor-manage.

This changes the fourth argument to be a flag, which if set, does
not display the tab.