Fix logged respirate sleep_duration_sec
Kernel#sleep returns the duration slept rounded to an integer, which
is less helpful when sleeps are going to be between 0.2 and 1.
Return the number of seconds that sleep was requested for instead,
which can be fractional. This should allow for proper calculation of
how long respirate scan threads are actually sleeping.