Clover: Support Ubiblk v0.1‑7
Ubiblk v0.1‑7 upgrades its dependencies to the most recent versions,
which consequently fixes a startup race condition.

The observable symptom of the race condition was that CH sometimes
failed to boot with Ubiblk disks. The sequence was:

1. CH sent `VHOST_USER_SET_VRING_ENABLE` to enable the vring, but didn’t
require an acknowledgement that it had been processed by Ubiblk.
2. The first IO request was placed into the vring, and the kick EventFd
was signalled to notify Ubiblk that there was IO to process.

The `vhost` crate used by Ubiblk ignores notifications before the vring
is enabled, so sometimes step 2 occurred before the enable message was
processed, causing Ubiblk to miss the IO request.

In vhost 0.14.0, the crate was updated to set the `REPLY_ACK` protocol
feature. When Ubiblk advertises this feature to CH, CH will wait for the
acknowledgement in step 1 before proceeding, eliminating the race
condition.