Add us-east-1 and us-east-2
Additionally, the amis will be added as part of the deployment.
Also, we had to tweak the bucket creation since us-east-1 is not a valid
parameter for location constraint. AWS guys are weird, if you don't pass
anything, the bucket is created in us-east-1, but the value us-east-1 is
not allowed. Such a world class API.
source: https://docs.aws.amazon.com/AmazonS3/latest/API/API_CreateBucketConfiguration.html