Install Metrics Server on new Kubernetes Clusters
This change adds a new step to Kubernetes cluster lifecycle to install
metrics server. The installation is done by applying a slightly modified
version the official yaml file [1] of the metrics server after the
cluster is provisioned. The metrics-server then installed on one of the
control plane nodes.

This works basically by incrementing a semaphore at the beginning of the
provisioning workflow and a label hop at the wait state of
KubernetesClusterNexus.

This approach can be extended to user-selectable addons and tools in the
future.

[1]
https://github.com/kubernetes-sigs/metrics-server/releases/latest/download/components.yaml