Add more plugins to CloverAdmin
This adds the render, exception_page (in development), public, flash,
and sessions plugin.

This adds a views/admin and public/admin directories to hold the
templates and public files for the admin site.

For the session secret, to avoid using the same session secret as
Clover itself, this uses an HMAC with the Clover session secret as
the key.