Support searching by ubid and navigating to associated objects in the admin site
For each object, this shows the object contents (other than redacted columns),
as well as links to all associated objects.

This adds the following plugins: typecast_params, symbol_matchers, not_found,
route_csrf, and forme_route_csrf.  Forme will be used to build the forms on
the admin site. Currently, there is only a single form, and it only does a
GET request and not a POST request, so the route_csrf code is not yet needed,
but it will be needed as soon as authentication is added.