Publish unique cache proxy log line counts to mezmo
To be able to follow problematic operations on cache proxy,
publishing the count for each unique log line. Each log line either
shows the succesful operations or the issues observed. Logging
their count with the line text in a map format will follow us to
make analysis on mezmo easier.