Check premium utilization for the premium concurrency check as well
Currently, we check the overall utilization to be below 75% to allow
exceeding the soft concurrency limit. However, this check is global and
doesn’t differentiate between standard and premium runners.
Consequently, if the standard utilization is low, but a single customer
with premium runners requests a large number of vCPUs, they could
consume all available premium capacity. As a result, other customers
with premium access might be blocked from getting premium runners.

If both standard and premium utilization are above 75%, we don’t allow
exceeding the soft concurrency limit.

If premium utilization is high, but standard utilization is low, we
allow exceeding the soft concurrency limit without a premium upgrade.

To accommodate requested premium customers, I reduced the limit to 50%
for free upgrades.