Add migrations for admin accounts
Admin site authentication will be separated from regular site
authentication. This adds tables to store admin accounts, password
hashes, and WebAuthn key information (admin site will require
password+WebAuthn authentication).

Similarly to normal site authentication, the password hash table
for the admin site will be stored in a table owned by the separate
database user, with database functions to get the salt and to
check for a valid password hash, but where the application database
user has no ability to read the password hashes.

This requires minor changes to the Rakefile, so that the second ph
migration is run.