Add Page.group_by_vm_host dataset method
This returns a hash of pages keyed by the VmHost ubid (or nil for
pages not related to a VmHost). To determine the related VMHost,
it looks at details["related_resources"] to find the related
objects, and tries to determine which VmHost the objects are
related to (if any).

I'm not sure that all "et" ubids should be considered ApiKeys
for this purpose, so this adds a keyword argument to
UBID.resolve_map to not make that assumption, and has
Page.group_by_vm_host use that keyword argument.