Fix use of locked domains
rodauth.omniauth_provider is a symbol, not a string. However, when using
OmniAuth.config.add_mock (the recommended way to mock with omniauth),
rodauth.omniauth_provider is a string, which is why the existing tests
did not catch this.  Use to_s so it will work regardless of whether
rodauth.omniauth_provider is a symbol or a string.