Add error handler to admin site
Without this, errors would be caught by Clover instead of CloverAdmin,
and probably would have been have looked bad, since it would try to
use the Clover stylesheet, and that would result in a 404.

This also changes the emit so we can differentiate admin route
exceptions from normal route exceptions.