Add handle_validation_failure web helper method
This method accepts a template, which will be rendered if there is
a validation failure, instead of issuing an internal request based
based on the referrer.  In the future, the method will accept a
block, but that isn't necessary yet.

This uses handle_validation_failure in the create project route to
show how it works. The plan is to expand the use of this method to
all POST web routes, and then drop the code that issues internal
requests. The internal request code has fewer issues than the code
it replaced (which did a browser redirect, and broke with large
forms), but it is still complex and brittle. In particular, storing
referrer information the flash breaks if a user has multiple tabs
open, and is getting form errors in one tab, and is using the other
tab to make changes.

One of the features of the internal request code, that dates back
to when redirects were used for errors, is the use of flash["old"]
to store parameters from the previous request.  This isn't
necessary if we aren't issuing an internal request, as we can look
directly in request.POST for the submitted parameters. This
changes the one component used for the project creation form, but
other templates accessing flash["old"] can be updated the same way.