Use handle_validation_failure in project name update route
Also, use a transaction around the project name update and the
audit logging.

Move the quota variable setting logic into the template to avoid
duplicating it.

Both the old and new capybara specs use the parameter name instead
of a label, because the text input does not have a label. Fixing
this requires modifications to the kv_data_card component.