Move the runner service setup to image generation
We begin to directly control the unit file instead of relying on
systemd-run’s transient unit generation at e2627b.

We tested it for a while in production, and it appears to be working
well. It’s now enabled for all customers.

Therefore, we can move it to image generation to reduce runtime
overhead.

https://github.com/ubicloud/runner-images/pull/37

Since the new image rollout takes some time, we retain the old code and
create these files if they don’t exist for backward compatibility.

I’ll remove it after the image rollout is complete.