Support monitor partitioning
This adds respirate-style partitioning to monitor, and refactors monitor.

With this, partitioning works the same way for monitor as it does
for respirate. There is a repartition thread that listens on a channel
for information about active partitions, and notifies about the current
process so that other monitor processes are aware of its existance.

As monitor is less time sensitive than respirate, only scanning for new
resources once a minute instead of once a second, the repartition thread
doesn't notify as often. Repartitioning for new monitor processes still
happens immediately, before the next enqueue (within 5 seconds). Removing
stale monitor processes happens within 1 minute.

In adding to adding partitioning, this refactors monitor significantly:

* This abstracts the monitor and metric export handling into a single
  Struct class.  Almost all of the logic was shared, and this removes
  the duplication.

* This avoids the separation of the scan thread from the main/enqueuing
  thread. Instead, the main/enqueing thread just scans if it has been
  more than a minute since the last scan, or if it was repartitioned since
  the last enqueue.  This removes the need for a mutex to protect the
  hash of resources.

* Previously, all resources that were initially monitored continued to
  be monitored until they were deleted. That doesn't work when
  repartitioning.  If the scan query does not pick up the resource, either
  because the resource was deleted or is no longer in the current partition,
  then it is no longer monitored.  It will keep the same resource previously
  used if it still exists, ensuring that the same SSH session is still used.

* Monitor now traps TERM (and INT), and generally exits cleanly within about
  1 second on shutdown unless worker threads are busy. It gives worker threads
  up to 2 seconds to exit, and if they don't exit within 2 second, the process
  exits with status 1, signifying failure.  If the process dies due to pulse
  checking or resource scanning failure, it exits with status 2.

This keeps the behavior of enqueuing locked resources, but I think that
behavior is questionable and should probably be removed.  It's better to
skip a locked resource in the current loop, and enqueue it in a later loop
after it is no longer locked.

In general, the enqueuing approach is questionable. We enqueue every monitored
resource every 5 seconds, regardless of the current status.  If it takes 4
seconds to run the job, then it is enqueued again 1 second after finishing.
We should probably have each resource keep track of when monitoring finished,
and only enqueue 5 seconds after the resource last completed its action. This
may result in increased overall time between monitoring in some cases, but
it can result in decreased overall time between monitoring in the case where
the resource was locked, assuming that we look at existing resources every
second, and only enqueue those that are not locked and finished over 5 seconds
ago.

The monitor process is now extensively documented. Hopefully, it should be
easier to understand.

We currently lack a test suite for monitor, which makes a change like this
risky. We should consider adding a test suite. We would need the equivalent
of Prog::Test for monitor. Then we could have the equivalent of the
respirate_smoke_test for monitor.