Outline mkdir_p for CSI directory initialization
It's unusual to have such a side effect upon object initialization.

In production cases, this object is not created more than once per
startup, so, no big deal...but once testing and mocks get involved, it
clutters the tests, so abide the more conventional approach of
outlining the side effect.

In passing, remove the check for existence of a directory before
`mkdir_p`: it idempotent, it will check if there's nothing to do
internally anyway.