Split monitor into testable classes
Move the MonitorResourceType class into it's own file.

Add MonitorRepartioner class to handle repartitioning.

Add MonitorRunner class to handle the main monitor loop:

* scan (if needed)
* report/emit metrics (if needed)
* check stuck pulses (if needed)
* enqueue

Each of the sections of the MonitorRunner main loop
are in separate methods, so they can be tested
independently.

This doesn't add tests, it only splits monitor into
the 3 classes. Tests for each class will be in subsequent
commits.

This fixes bugs found when testing.