Use HTTP header instead of meta tag for OTP lockout refresh
Hopefully the HTTP header is more reliable than the meta tag. The
meta tag was previously placed in the body and not the header,
which isn't standard, so it's possible that's why some browsers
didn't respect it.

Additionally, show the remaining time to complete authentication
in seconds instead of a timestamp that may not match the user's
timezone. Also show the number of seconds until the refresh.

This requires a new feature in Rodauth, so update to the latest
Rodauth commit.