Add a little LogDNA aka Mezmo client
For interactive terminals, we want to take control over how we send
logs for auditing, it cannot be delegated to a some kind of
supervisory process unless we wanted to get all the terminal control
sequences as well.  In many cases, we don't want that, but do want to
record, for example, what ruby code was evaluated in `pry`.

Alas, I couldn't find a convenient lower common denominator transport
for sending logs, e.g. ndjson or syslog+tls, so vendor-specific
support is what I implement here.

There is a LogDNA gem, but it's a hassle to make it optional for
development and yet installed for production.  Having a little
internal client will allow us to ignore a logging implementation if
some environment variables are not set without bringing bundler and
rubygems into the mix.