Increase ssh keep_alive interval to 60 seconds
As our load increased, we scaled up accordingly, and the current
keep_alive interval, which is set to 3 seconds, is too aggressive.
By increasing the interval, we allow the VmHosts to catch up with
the keepalive packets.

Right now, we are getting many SSH exceptions in LoadBalancerVmPort
monitor like this:
class IOError, message: closed stream
for loadbalancer healthchecks.